add image processing
This is memory efficient enough to operate on 240x208 pixel images.
(An earlier iteration which used float32s was not)

Typically an algorithm takes 1 to 4 seconds to run on an image. Channel
operations such as solarize are faster, while convolution operations like
sharpen are slower.

A range of algorithms are provided and there are building blocks to create
others.